Wav2Lip-former/wav2lip/requirements.txt
librosa==0.7.0
opencv-contrib-python>=4.2.0.34
opencv-python==4.1.0.25
tqdm==4.45.0
numba==0.48
pydub
baidu_aip
chardet